src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-cp-core>=1.0.0,<1.1.0
python-libmaas==0.6.5


cloudshell-shell-standards>=1.0,<2.0
cloudshell-shell-flows>=1.0,<2
cloudshell-automation-api>=9.2
